scripts/requirements.txt
azure-cognitiveservices-speech==1.28.0
pydub==0.25.1
